Rename Dtos to Entities
In the project, data transfer objects are being used as data models.
I thought it is because to keep the test case as simple as possible.
Then I noticed the project also contains a Customer class which is
used as Dto. It creates confusion. I moved all data models into the
Entities folder and move the Dtos into the Dtos folder to make it more
explicit and intuitive.
Please see https://en.wikipedia.org/wiki/Data_transfer_object